🚚 move `NavLink` component into its own file
